General code clean up.
After the initial bout of getting this thing running,
there were a lot of areas to refine code.

Also improved performance in some respects by continuing
if the url_substring is malformed before iterating
through all of the visited URLs.

Changed `get_url_components` to handle the
`MalformedURLException` instead of the caller.  And also
used a single try-catch for joining the threads.